pyproject.toml
[tool.poetry]
name = "junior-jobs-api"
version = "0.1.0"
description = "API for junior-jobs-bot"
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"
packages = [
    {include = "api"}
]

[tool.poetry.dependencies]
python = "^3.10.8"
fastapi = "^0.88.0"
sqlalchemy = "^1.4.44"
python-dotenv = "^0.21.0"
uvicorn = "^0.20.0"
psycopg2 = "^2.9.5"
requests = "^2.28.1"
loguru = "^0.6.0"
fastapi-pagination = "^0.11.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
loguru = "^0.6.0"
pytest = "^7.2.0"
httpx = "^0.23.1"
pytest-cov = "^4.0.0"
isort = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
